Auto-reconnect bug fixed.
There was a stupid bug that was stopping bfirc from rejoining channels
when it reconnected after a disconnect - it was checking if the channel
was already in the buffers, which is stupid, as the join stuff is event-
driven anyway (i.e. it does all the join stuff on the client side when
the server tells it that it's joined the user into a channel) so it
shouldn't break anything by trying to join a channel twice.