Reconnect stuff improved.
The reconnect stuff was causing lots of unnecessary output
to be sent to all the buffers and the last commit should have
fixed it but there was still a bit too much cunt involved.

The (server) buffer will still fill up with "Reconnecting
in 30 seconds..." so over an hour or so that would be
quite a lot but at least it's just one screen. I'm thinking
of a good way to have it just display one line for the entire
disconnect but update it. I think just showing the time of
disconnect and then having a single "Reconnecting in..." line
perhaps with the number of attemps or something. Either way,
I think at the moment it's less critical than it was.