multi-connection support seems to be working
I've tested it a bit and it seems to be doing what you'd generally expect it to but there are issues that I know of; pathalogical cases that will cause things to really fuck up. As far as I can tell it's perfectly usable, but you'll have to keep an eye in the (server) window for commands that would depend on the currently selected connection, e.g. joining channels. Normal stuff like saying stuff in a channel, /me actions, etc. should do what you'd expect automatically but I'm sure I haven't taken care of all situations.

Basically this is the inital implementation for multi connection support so I've made a new branch for it (multicon). I'm assuming I'll be able to push forward with this but any real world experience with the pathological cases (I'm not going to encourage you to fuck it up by telling you what they are; it's not hard to work out anyway probably) should be reported back to me so I can think of an appropriate and tidy way of fixing it. Enjoy.