Lots more work on the multicon stuff plus a stupid character stripping bug fixed
I think the multicon stuff is in place and complete to an extent; the real trouble
will come along when the program has to 'guess' which network the user wants a
command to go to. So feedback from other people will hopefully let me streamline
that.

There was also a bug that would strip either n or o from the end of lines, which
was stupid and is fixed now; it was part of the solution to stop superfluous
\n's being in messages if you paste them in.